Refactor(event): 행사 게시판, 게시글 컴포넌트 구조 수정
