--- src/app/(main)/event/board/_components/EventBoardHero/EventBoardDetail/index.tsx
@@ -1,13 +1,0 @@
-export const EventBoardDetail = () => {
-  return (
-    <div className="flex flex-col gap-2 pb-5 pl-1">
-      <div className="text-2xl font-semibold">행사 게시판</div>
-      <div className="flex flex-col gap-1 text-sm">
-        <div>🗓️ 개강총회/종강총회/성과공유회</div>
-        <div>🎉 해크닉/MT/슬기로운 해달생활</div>
-        <div>💡 아이디어톤/해커톤</div>
-        <div>🎓 진로특강/취업특강</div>
-      </div>
-    </div>
-  )
-}
