+++ src/app/(main)/event/board/_components/create-post-button/CreatePostButton.tsx
