+++ src/app/(main)/event/board/_components/create-post-button/index.ts
@@ -1,0 +1,1 @@
+export { CreateEventPostButton } from './CreatePostButton'
