+++ src/app/(main)/event/board/_components/hero/index.ts
@@ -1,0 +1,1 @@
+export { EventBoardHero } from './Hero'
