+++ src/app/(main)/event/board/_components/index.ts
@@ -1,0 +1,3 @@
+export * from './hero'
+export * from './post-list'
+export * from './create-post-button'
