+++ src/app/(main)/event/board/_components/post-list/index.ts
@@ -1,0 +1,1 @@
+export { EventPostList } from './PostList'
