+++ src/app/(main)/event/board/create-post/_components/form/index.ts
@@ -1,0 +1,1 @@
+export { CreateEventPostForm } from './Form'
