+++ src/app/(main)/event/board/create-post/_components/hero/Hero.tsx
