+++ src/app/(main)/event/board/create-post/_components/hero/index.ts
@@ -1,0 +1,1 @@
+export { CreateEventPostHero } from './Hero'
