+++ src/app/(main)/event/board/create-post/_components/index.ts
@@ -1,0 +1,2 @@
+export * from './form'
+export * from './hero'
