+++ src/app/(main)/event/board/posts/[postId]/_components/hero/Hero.tsx
