+++ src/app/(main)/event/board/posts/[postId]/_components/hero/index.ts
@@ -1,0 +1,1 @@
+export { EventPostHero } from './Hero'
