+++ src/app/(main)/event/board/posts/[postId]/_components/index.ts
@@ -1,0 +1,2 @@
+export * from './hero'
+export * from './post-detail'
