+++ src/app/(main)/event/board/posts/[postId]/_components/post-detail/PostDetail.tsx
