+++ src/app/(main)/event/board/posts/[postId]/_components/post-detail/index.ts
@@ -1,0 +1,1 @@
+export { EventPostDetail } from './PostDetail'
