+++ src/app/(main)/event/board/posts/[postId]/page.tsx
@@ -2,8 +2,11 @@
 
 import { Suspense } from 'react'
 
-import { EventPostHero } from './_components/EventPostHero'
-import { EventPostSection } from './_components/EventPostSection'
+import { Spinner } from '@/components/common'
+import { BoardNavigationButton, PostContent } from '@/components/feature'
+import { useGetPost } from '@/service/data/post'
+
+import { EventPostDetail, EventPostHero } from './_components'
 
 type EventPostPageParams = {
   params: {
@@ -12,11 +15,19 @@ type EventPostPageParams = {
 }
 
 const EventPostPage = ({ params }: EventPostPageParams) => {
+  const { data: post, status } = useGetPost({ postId: Number(params.postId) })
+
+  if (status === 'pending') return <Spinner />
+
+  if (!post) return <div>게시글 정보가 없습니다.</div>
+
   return (
     <div>
       <EventPostHero />
       <Suspense fallback={<div>loading...</div>}>
-        <EventPostSection postId={Number(params.postId)} />
+        <EventPostDetail post={post} />
+        <PostContent content={post.postContent} />
+        <BoardNavigationButton />
       </Suspense>
     </div>
   )
